backend/requirements.txt
# =====================================================
# FASTAPI STACK
# =====================================================

fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart
email-validator

# =====================================================
# DATABASE
# =====================================================

sqlalchemy>=2.0.37,<2.1
psycopg2-binary
alembic

# =====================================================
# SECURITY
# =====================================================

python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# =====================================================
# ENV CONFIG
# =====================================================

python-dotenv
pydantic-settings

# =====================================================
# PAYMENTS
# =====================================================

razorpay==1.4.2

# =====================================================
# AI / OPENAI
# =====================================================

openai>=1.30.0

# =====================================================
# HTML + PDF GENERATION
# =====================================================

jinja2
playwright
pillow

# =====================================================
# CHARTS / ANALYTICS
# =====================================================

matplotlib
numpy

# =====================================================
# OPTIONAL UTILS
# =====================================================

python-dateutil
pypdf
# Razorpay depends on pkg_resources, removed in Setuptools >=81.
setuptools<81
